* origin:
  chore(deps-dev): bump webdriverio from 7.9.1 to 7.10.0 (#3001)
  chore(deps): bump flyway-core from 7.13.0 to 7.14.0 (#3002)
  feat: Resolve cases per child with different reasons (FPLA-2855) (#2994)
  chore: add more endpoints to audit.json (confirmed with Tech Lead this alert type was agreed with Security team to be safe to ignore) (#3003)
  Onboard isle of wight (#2998)
  Adding change to respondent solicitor (#2992)
  FPLA-2869 Trigger Welsh translation for orders (#2973)
  chore(deps-dev): bump webdriverio from 7.8.0 to 7.9.1 (#2981)
  release: toggle on child solicitor (FPLA-3306) (#2993)
  fix: Fix Others event to save non-empty others (FPLA-3303) (#2980)
  doc: add welsh translation of FL-PLW-LET-ENG-00232.doc (#2997)
  Fpla 3254 Trigger Welsh translations for SDO (#2962)
  chore(deps): bump flyway-core from 7.12.0 to 7.13.0 (#2988)
  chore: update xui config (#2989)
  temp: resture prev xui version until pernamet fix is found (#2991)
  chore: Remove fpl service token from integrtion tests Constants (#2990)

+++ service/src/main/java/uk/gov/hmcts/reform/fpl/model/docmosis/welshtranslation/DocmosisTranslationRequest.java
@@ -19,7 +19,7 @@ public class DocmosisTranslationRequest implements DocmosisData {
     private String description;
     private DocmosisTranslateLanguages translate;
     private DocmosisWelshLayout layout;
-    private int wordCount;
+    private long wordCount;
     private String dateOfReturn;
     private RenderFormat format;
 }
